analytics/requirements.txt
mysql-connector-python
pymongo
apscheduler